+++ goodreadskosync.koplugin/_meta.lua
@@ -1,5 +1,5 @@
 local _ = require("gettext")
 return {
-    fullname = _("Goodreads Sync (unofficial)"),
+    fullname = _("Goodreads KO Sync"),
     description = _("Keep your KOReader reading in sync with Goodreads."),
 }